add ChatWindow and ChatWindowHeader components; restructure the ChatBox page;
